:art: refactor : AddCardCommand enum class의 위치를 View에서 Domain으로 이동
