+++ src/main/java/domain/player/Player.java
@@ -5,7 +5,7 @@ import domain.card.Deck;
 import domain.card.Score;
 import domain.player.state.Ready;
 import domain.player.state.State;
-import view.AddCardCommand;
+import domain.game.AddCardCommand;
 
 import java.util.List;
 import java.util.Objects;
